todo_backend/todo_app/serializers.py
# TodoSerializer is configured to serialize the 5 fields from the Todo model

from .models import Todo
from rest_framework import serializers

class TodoSerializer(serializers.ModelSerializer):
    class Meta:
        model = Todo
        fields = ['id', 'title', 'description', 'due_date', 'status']